requirements.txt
faker==37.11.0
pandas==2.1.4
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
colorlog==6.9.0

googlemaps==4.10.0

#SQLAlchemy==2.0.43
SQLAlchemy==1.4.49
psycopg2-binary==2.9.9

apache-airflow[postgres]==2.10.3 \
    --constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.10.3/constraints-3.11.txt"
